Deals with Security methods and ascii art.
It contains a class Security that deals with encryption and decryption using fernet module. It also contains 3 methods:

1. setPrimaryKey() : sets new fernet key or symmetric key to file key.pem
2. encrypt_json() : encrypts the data.json
3. decrypt_json() : decrypts the data.json

This module is used by the main pwdManager.py. If this module is run individually then it shows the ascii art and calls the setPrimaryKey() and encrypts the data.json for the first time using new key